refactor(security): relocate sanitizeForCell to new safe-writes module
