+++ src/server/index.ts
@@ -33,9 +33,9 @@ import {
   interpolateTemplate,
   flattenArg,
   markAIOutputRange,
-  sanitizeForCell,
   resolveGroundingUris,
 } from "./utils";
+import { sanitizeForCell } from "./safe-writes";
 import { CONFIG } from "./config";
 import type {
   RunConfig,
